% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inheritance.R
\name{data_inheritance}
\alias{data_inheritance}
\title{Prepare inheritance data}
\usage{
data_inheritance(
  path_survey = "./inst/dataINSEE/Enquete Patrimoine",
  year = 2015
)
}
\arguments{
\item{path_survey}{Path to survey data}

\item{year}{Year}
}
\description{
Prepare inheritance data
}
