% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_inheritance.R
\name{ordered_model_threshold}
\alias{ordered_model_threshold}
\title{Function to estimate polytomic discrete order models}
\usage{
ordered_model_threshold(
  data,
  formula = "MTHER ~ tr_agfinetu + tr_age + revenu + SEXE",
  formulaSD = NULL,
  link = c("logit", "probit"),
  delta = NULL,
  constantMEAN = TRUE,
  constantSD = TRUE,
  thresholds = NULL,
  ...
)
}
\arguments{
\item{data}{Dataframe used for estimation}

\item{formula}{An object of class \link[stats]{formula}:
a symbolic description of the model used to explain the
mean of the latent variable. The response variable should
be a numeric vector or factor variable such that the
numerical assignments for the levels of the
factor have ordinal meaning.}

\item{formulaSD}{either \code{NULL} (homoskedastic model) or an object
of class \link[stats]{formula}: a symbolic description of
the model used to explain the variance of the latent variable.}

\item{link}{specifies a link function for the model to be
estimated, accepted values are \emph{"probit"},
\emph{"logit"}, \emph{"cauchit"}, \emph{"loglog"} and \emph{"cloglog"}}

\item{delta}{\code{NULL} or numeric vector. Used to prespecify
elements of the parameter vector for the equation of
the variance of the latent variable. Vector should be of
length one or of length equal to the number of explanatory
variables in the variance equation.
If of length one the value is presumed to
correspond to the constant if a constant is included
or the first element of the parameter vector. If
of length greater than one then \code{NA} should be entered
for elements of the vector to be estimated.}

\item{constantMEAN}{logical. Should an intercept be
included in the model of the mean of the latent variable?
Can be overwritten and set to \code{FALSE} using the
\code{formulaMEAN} argument by writing \verb{0 +}
as the first element of the equation.}

\item{constantSD}{logical. Should an intercept be included
in the model of the variance of the latent variable?
Can be overwritten and set to \code{FALSE} using the \code{formulaSD}
argument by writing \verb{0 +} as the first element of the equation.}

\item{thresholds}{\code{NULL} (default) or numeric vector.
Thresholds that define discretization
of the response variable. Must either be a numeric vector,
assuming a known threshold ordered discrete model or
\code{NULL} value in which case a unordered discrete
choice model is applied. Vector should be of length
equal to the number of outcomes
minus one. \code{NA} should be entered for threshold
parameters to be estimated by the model.}

\item{...}{Additional parameters that should be supplied
to \link[oglmx]{oglmx} function}
}
\description{
Function to estimate polytomic discrete order models
}
\details{
Behind the stage, \link[oglmx]{oglmx} is
 used to estimate models
 for which the outcome variable is discrete and the mean
 and/or variance of the underlying latent variable can
 be modelled as a linear combination of explanatory
 variables. \link[oglmx]{oglmx} is preferred to
 \link[MASS]{polr} because the latter does not
 allpw constrained models where thresholds in latent
 space for response variable are known
}
\examples{
iris$y <- sample(1:5, size = nrow(iris),
                replace = TRUE)
summary(
 ordered_model_threshold(
  iris,
  formula = "y ~ Sepal.Length + Sepal.Width + Petal.Length + Petal.Width",
  link = "probit",
  thresholds = NULL)
)

}
\seealso{
\link[oglmx]{oglmx}, \link[MASS]{polr}
}
