% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_inheritance.R
\name{predict_inheritance}
\alias{predict_inheritance}
\title{Predict inheritance from an econometric model}
\usage{
predict_inheritance(
  inheritance_model,
  data,
  selection_model = NULL,
  scale = "log",
  bias_correction = FALSE,
  variable_name = "y_pred",
  lbounds = NULL
)
}
\arguments{
\item{inheritance_model}{A `oglmx` model}

\item{data}{Dataframe}

\item{scale}{Scale in the latent variable space}

\item{bias_correction}{Logical value indicating, if the
model is in `log` whether predictions should be corrected
by a factor \eqn{exp(-\sigma^2/2)}}

\item{variable_name}{Variable name for the prediction}
}
\value{
Initial dataframe with a new column corresponding to
 `variable_name` supplied value
}
\description{
Predict inheritance from an econometric model
}
