% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_estimation.R
\name{prepare_estimation}
\alias{prepare_estimation}
\title{Prepare data for estimation}
\usage{
prepare_estimation(
  data_estimation,
  .cols = c("MTHER", "AGFINETU", "AGE", "ZSALAIRES_I", "ZRETRAITES_I", "ZCHOMAGE_I",
    "SEXE"),
  taille_tr_age = 5,
  taille_tr_agfinetu = 2
)
}
\arguments{
\item{.cols}{Columns to keep}

\item{data}{Wealth survey data from \link{data_inheritance}}
}
\description{
Prepare data for estimation
}
