% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inheritance.R
\name{read_inheritance}
\alias{read_inheritance}
\title{Read inheritance data}
\usage{
read_inheritance(
  path_survey = "./inst/dataINSEE/Enquete Patrimoine",
  year = 2009
)
}
\arguments{
\item{path_survey}{Path to survey data}

\item{year}{Year}
}
\description{
Read inheritance data
}
