% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{taxation_inheritance}
\alias{taxation_inheritance}
\title{From observed inheritance, determine
 inheritance that is been given}
\usage{
taxation_inheritance(x, deduction, bounds_legal, taxes)
}
\arguments{
\item{x}{Observed inheritance vector}

\item{deduction}{Level of tax deduction}

\item{bounds_legal}{Bounds for the inheritance
tax rates}

\item{taxes}{Tax rates to be applied}
}
\description{
From observed inheritance, determine
 inheritance that is been given
}
